pyproject.toml
[project]
name = "logger-bot-notifier"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "MystiSs", email = "vanz.vaf@yandex.ru"},
]
dependencies = ["python-telegram-bot>=22.5", "fastapi>=0.119.0", "uvicorn>=0.37.0", "pydantic-settings>=2.11.0", "aiohttp>=3.13.0"]
requires-python = "==3.13.7"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
run-server = "python src/server.py"
run-error-test = "python src/tests/test_error_send.py"
